Moved headers into their own folder and updated makefile (currently not building correctly)
