+++ sourcecode/Part1/include/commandfunctions.h
